feat(extensable identifier): moved the geration of the namespace identifier and if the class should be added to a mixin for easier exensability
