+++ addon/mixins/style-namespacing-extras.js
@@ -1,0 +1,21 @@
+import Ember from 'ember';
+
+const {
+  Component,
+  computed,
+  Mixin,
+} = Ember;
+
+export default Mixin.create({
+  _componentIdentifier: computed({
+    get() {
+      return (this._debugContainerKey || '').replace('component:', '');
+    }
+  }),
+
+  _shouldAddNamespacedClassName: computed({
+    get() {
+      return this.get('tagName') !== '' && this.get('componentCssClassName');
+    }
+  }),
+});
