+++ app/mixins/style-namespacing-extras.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-component-css/mixins/style-namespacing-extras';
